Fetch car images
In this update, the images of all car makes and models are fetched
from image-studio API.
Also, whenever a car model is searched, we do not make a new API
call to the server. Instead, we change the URL parameters based on
the values in our search bar and NextJs automatically
refetches the data due to its server side rendering ability.