Add custom filters
This update adds custom filters for filtering our cars.
Again, we use the server-side rendering ability of NextJs to filter
our cars anytime the filter value changes- We will not be making
new API calls to the server.
The update also adds a showMore button to show more cars if needed.